requirements.txt
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
openpyxl==3.1.2
pyarrow==15.0.0